Revert "refactor: LikeController 좋아요 기능 로직 개선 - Repository 반환 타입 Optional로 변경 - LikeStatusResponse 클래스 Data 어노테이션 제거 - 좋아요 클릭 기능 내부 로직, 빠른 반환하도록 개선"
This reverts commit d8db62cacceae1c0033765feef5a2b23079bfefe.

+++ src/main/java/com/bugflix/weblog/like/service/LikeServiceImpl.java
@@ -37,18 +37,21 @@ public class LikeServiceImpl {
      * <p>
      * - isLiked 와 likeCount 를 LikeStatusResponse 로 encapsulation 하여 반환
      */
-    public LikeStatusResponse changeLikeStatus(Long postId, UserDetails userDetails) throws Exception {
+    public LikeStatusResponse changeLikeStatus(Long postId, UserDetails userDetails) {
+        LikeStatusResponse likeStatus = new LikeStatusResponse();
 
         Long userId = ((CustomUserDetails)userDetails).getUser().getUserId();
 
-        Like like = likeRepository.findLikeById_PostIdAndId_UserId(postId, userId).orElseThrow(Exception::new);
+        Like like = likeRepository.findLikeById_PostIdAndId_UserId(postId, userId);
 
-        if (like != null) {
-            likeRepository.delete(like);      // 본인의 Like 상태 변경 Logic
-            return new LikeStatusResponse(likeRepository.countByIdPostId(postId), false);
+        if (like != null) likeRepository.delete(like);      // 본인의 Like 상태 변경 Logic
+        else {
+            likeRepository.save(new Like(userId, postId));
+            likeStatus.setLiked(true);
         }
-        likeRepository.save(new Like(userId, postId));
-        return new LikeStatusResponse(likeRepository.countByIdPostId(postId), true);
 
+        likeStatus.setLikeCount(likeRepository.countByIdPostId(postId));    // Like Count 변경 Logic
+
+        return likeStatus;
     }
 }
